+++ site/admin/pages/mainpage.php
@@ -10,6 +10,7 @@
     <a href="<?php echo $_SERVER['PHP_SELF']; ?>?URL=useredit"><?php echo $U->getLang("admin.user.edit") ?></a><br />
     <a href="<?php echo $_SERVER['PHP_SELF']; ?>?URL=editor"><?php echo $U->getLang("admin.edit.new.site") ?></a><br />
     <a href="<?php echo $_SERVER['PHP_SELF']; ?>?URL=blogeditor"><?php echo $U->getLang("admin.edit.new.blogsite") ?></a><br />
+    <a href="<?php echo $_SERVER['PHP_SELF']; ?>?URL=deletepage"><?php echo $U->getLang("admin.delete") ?></a><br />
     <a href="<?php echo $_SERVER['PHP_SELF']; ?>?URL=about"><?php echo $U->getLang("admin.about") ?></a><br />
     <form action="<?php echo $_SERVER['PHP_SELF']; ?>">
       <select name="SiteName">
@@ -21,9 +22,10 @@
             echo "<option value='".$zeile["Name"]."'>".$zeile["Name"]."</option>";
           }
         ?>
-     </select>
-    <input type="hidden" name="URL" value="editor" />
-      <button type="submit"><?php echo $U->getLang("admin.edit.site") ?></button></form>
+      </select>
+      <input type="hidden" name="URL" value="editor" />
+      <button type="submit"><?php echo $U->getLang("admin.edit.site") ?></button>
+    </form>
       <form action="<?php echo $_SERVER['PHP_SELF']; ?>">
         <select name="SiteName">
           <?php